Move executeConsistentStateUpdate method from Repository to SnapshotsService
This refactoring helps make https://github.com/elastic/elasticsearch/pull/90332 easier to read.
Also, even in isolation it saves a lot of code and seems to generally be the right thing to have this
method in the snapshots service.
